@@ -1,26 +1,26 @@
 package com.levimoreira.teammenagerapp.home
 
-
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.levimoreira.teammenagerapp.R
+import com.levimoreira.teammenagerapp.databinding.FragmentHomeBinding
+import dagger.hilt.android.AndroidEntryPoint
 
-/**
- * A simple [Fragment] subclass.
- *
- */
+@AndroidEntryPoint
+class HomeFragment : Fragment(R.layout.fragment_home) {
+    private var _binding: FragmentHomeBinding? = null
+    private val binding get() = _binding!!
 
-class HomeFragment : Fragment() {
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        _binding = FragmentHomeBinding.bind(view)
     }
 
-    companion object {
-        val TAG = "HomeFragment"
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
     }
 }
