@@ -1,18 +1,19 @@
 package com.levimoreira.teammenagerapp.person.adapters
 
-import android.support.v7.widget.RecyclerView
-import android.view.View
+import android.view.LayoutInflater
+import androidx.recyclerview.widget.RecyclerView
 import android.view.ViewGroup
-import com.levimoreira.teammenagerapp.R
 import com.levimoreira.teammenagerapp.application.entities.Person
-import com.levimoreira.teammenagerapp.inflate
-import kotlinx.android.synthetic.main.person_item_view.view.*
+import com.levimoreira.teammenagerapp.databinding.PersonItemViewBinding
 
 class PersonListAdapter(val list: List<Person>) : RecyclerView.Adapter<PersonListAdapter.PersonViewHolder>() {
 
     lateinit var clickListener: (Person) -> (Unit)
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PersonViewHolder = PersonViewHolder(parent.inflate(R.layout.person_item_view))
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PersonViewHolder {
+        val binding = PersonItemViewBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        return PersonViewHolder(binding)
+    }
 
     override fun getItemCount(): Int = list.size
 
@@ -21,12 +22,15 @@ class PersonListAdapter(val list: List<Person>) : RecyclerView.Adapter<PersonLis
         holder.bind(person, this.clickListener)
     }
 
-    inner class PersonViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-        fun bind(item: Person, clicklistener: (Person) -> Unit) = with(itemView) {
-            name.text = item.name
-            phone.text = item.phone
-            email.text = item.email
-            setOnClickListener { clicklistener(item) }
+    inner class PersonViewHolder(private val binding: PersonItemViewBinding) : RecyclerView.ViewHolder(binding.root) {
+
+        fun bind(item: Person, clicklistener: (Person) -> Unit) = with(binding.root) {
+            binding.apply {
+                name.text = item.name
+                phone.text = item.phone
+                email.text = item.email
+                setOnClickListener { clicklistener(item) }
+            }
         }
     }
 }
