@@ -1,28 +1,27 @@
-package com.levimoreira.teammenagerapp.application.di
-
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
-import javax.inject.Inject
-import javax.inject.Provider
-import javax.inject.Singleton
-
-@Suppress("UNCHECKED_CAST")
-@Singleton
-class DaggerViewModelFactory
-@Inject constructor(
-        private val creators: Map<Class<out ViewModel>, @JvmSuppressWildcards Provider<ViewModel>>
-) : ViewModelProvider.Factory {
-
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        val creator = creators[modelClass]
-                ?: creators.asIterable().firstOrNull { modelClass.isAssignableFrom(it.key) }?.value
-                ?: throw IllegalArgumentException("unknown model class $modelClass")
-
-        return try {
-            creator.get() as T
-        } catch (e: Exception) {
-            throw RuntimeException(e)
-        }
-
-    }
-}
+//package com.levimoreira.teammenagerapp.application.di
+//
+//import androidx.lifecycle.ViewModel
+//import androidx.lifecycle.ViewModelProvider
+//import javax.inject.Inject
+//import javax.inject.Provider
+//import javax.inject.Singleton
+//
+//@Suppress("UNCHECKED_CAST")
+//@Singleton
+//class DaggerViewModelFactory
+//@Inject constructor(a
+//    private val creators: Map<Class<out ViewModel>, @JvmSuppressWildcards Provider<ViewModel>>
+//) : ViewModelProvider.Factory {
+//
+//    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+//        val creator = creators[modelClass]
+//            ?: creators.asIterable().firstOrNull { modelClass.isAssignableFrom(it.key) }?.value
+//            ?: throw IllegalArgumentException("unknown model class $modelClass")
+//
+//        return try {
+//            creator.get() as T
+//        } catch (e: Exception) {
+//            throw RuntimeException(e)
+//        }
+//    }
+//}
