@@ -1,11 +1,11 @@
-package com.levimoreira.teammenagerapp.application.di
-
-import android.arch.lifecycle.ViewModelProvider
-import dagger.Binds
-import dagger.Module
-
-@Module
-abstract class DaggerViewModelFactoryModule {
-    @Binds
-    internal abstract fun bindViewModelFactory(factory: DaggerViewModelFactory): ViewModelProvider.Factory
-}
+//package com.levimoreira.teammenagerapp.application.di
+//
+//import androidx.lifecycle.ViewModelProvider
+//import dagger.Binds
+//import dagger.Module
+//
+//@Module
+//abstract class DaggerViewModelFactoryModule {
+//    @Binds
+//    internal abstract fun bindViewModelFactory(factory: DaggerViewModelFactory): ViewModelProvider.Factory
+//}
