@@ -1,9 +1,9 @@
-package com.levimoreira.teammenagerapp.application.di
-
-import android.arch.lifecycle.ViewModel
-import dagger.MapKey
-import kotlin.reflect.KClass
-
-@Target(AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY_GETTER, AnnotationTarget.PROPERTY_SETTER)
-@MapKey
-internal annotation class ViewModelKey(val value: KClass<out ViewModel>)
+//package com.levimoreira.teammenagerapp.application.di
+//
+//import androidx.lifecycle.ViewModel
+//import dagger.MapKey
+//import kotlin.reflect.KClass
+//
+//@Target(AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY_GETTER, AnnotationTarget.PROPERTY_SETTER)
+//@MapKey
+//internal annotation class ViewModelKey(val value: KClass<out ViewModel>)
