@@ -3,27 +3,27 @@ package com.levimoreira.teammenagerapp.business.data
 import com.levimoreira.teammenagerapp.application.entities.Business
 import io.reactivex.Maybe
 import io.reactivex.Single
+import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class BusinessRepository @Inject constructor(private var dao: BusinessDao) {
-    fun getBusinessList(): Single<List<Business>> {
+
+    fun getBusinessList(): Flow<List<Business>> {
         return dao.getAll()
     }
 
-    fun getSingleBusiness(id: Long): Maybe<Business> {
-        return dao.getOrganizationById(id)
+    fun getSingleBusiness(id: Long): Flow<Business> {
+        return dao.getBusinessById(id)
     }
 
-    fun insertBusiness(business: Business): Single<Long> {
-        return Single.fromCallable {
-            dao.insert(business)
-        }
+    suspend fun insertBusiness(business: Business): Long {
+        return dao.insert(business)
+
     }
 
-    fun deleteBusiness(id: Long): Single<Int> {
-        return Single.fromCallable {
-            dao.deleteById(id)
-        }
+    suspend fun deleteBusiness(id: Long): Int{
+        return dao.deleteById(id)
+
     }
 
 
