@@ -1,25 +1,26 @@
 package com.levimoreira.teammenagerapp.organization.data
 
-import android.arch.persistence.room.Dao
-import android.arch.persistence.room.Insert
-import android.arch.persistence.room.OnConflictStrategy
-import android.arch.persistence.room.Query
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
 import com.levimoreira.teammenagerapp.application.entities.Organization
 import com.levimoreira.teammenagerapp.application.entities.Person
 import io.reactivex.Maybe
 import io.reactivex.Single
+import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface OrganizationDao {
     @Query("SELECT * from " + Organization.TABLE_NAME)
-    fun getAll(): Single<List<Organization>>
+    fun getAll(): Flow<List<Organization>>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun insert(org: Organization): Long
+    suspend fun insert(org: Organization): Long
 
     @Query("SELECT * FROM " + Organization.TABLE_NAME + " WHERE id = :orgId")
-    fun getOrganizationById(orgId: Long): Maybe<Organization>
+    fun getOrganizationById(orgId: Long): Flow<Organization>
 
     @Query("DELETE FROM " + Organization.TABLE_NAME + " WHERE id = :org")
-    fun deleteById(org: Long): Int
+    suspend fun deleteById(org: Long): Int
 }
