@@ -1,23 +1,16 @@
 package com.levimoreira.teammenagerapp.person.viewmodel
 
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.LiveDataReactiveStreams
-import android.arch.lifecycle.ViewModel
-import com.levimoreira.teammenagerapp.application.entities.Person
+import androidx.lifecycle.SavedStateHandle
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.asLiveData
 import com.levimoreira.teammenagerapp.person.data.PersonRepository
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
+import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
-class PersonListViewModel @Inject constructor(var personRepository: PersonRepository) : ViewModel() {
+@HiltViewModel
+class PersonListViewModel @Inject constructor(private val personRepository: PersonRepository) :
+    ViewModel() {
 
-    fun getAllPersons(): LiveData<List<Person>> {
-        val result = personRepository
-                .getPersonList()
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .toFlowable()
-
-        return LiveDataReactiveStreams.fromPublisher(result)
-    }
+    private var _allPersons = personRepository.getPersonsList().asLiveData()
+    val allPersons = _allPersons
 }
